fix(api): refactor import & export code for pipe
