+++ apps/api/src/app/shared/index.ts
@@ -1,3 +1,3 @@
 export * from './shared.module';
-export { UUIDValidationPipe } from './pipes/uuid-validation.pipe';
+export * from './pipes';
 export * from './handlers';
