+++ apps/api/src/app/shared/pipes/index.ts
@@ -1,0 +1,2 @@
+export * from './uuid-validation.pipe';
+export * from './parse-json.pipe';
